requirements.txt
Flask
SkPy
openai
python-dotenv
flask-cors
gunicorn